# crates/secure_smoke_service/openapi.yaml
openapi: "3.1.0"
info:
  title: SunLit Security Smoke-Test Service
  description: >
    A purpose-built axum microservice with 54 routes, each exercising a
    specific security control against a known attack class. Includes 15
    mobile security routes covering MASVS controls. Designed for automated
    DAST scanning with OWASP ZAP.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:3001
    description: Local development server
paths:
...
  /smoke/xss:
    post:
      summary: XSS payload is HTML-encoded on reflection
      operationId: smokeXss
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XssRequest"
      responses:
        "200":
          description: HTML-encoded reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XssResponse"
  /smoke/sqli:
    post:
      summary: SQL-injection payload rejected via SqlIdentifier
      operationId: smokeSqli
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SqliRequest"
      responses:
        "200":
          description: Valid SQL identifier accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SqliResponse"
        "422":
          description: Invalid SQL identifier rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /smoke/cmdi:
    post:
      summary: Command-injection payload rejected via SafeCommandArg
      operationId: smokeCmdi
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmdiRequest"
      responses:
        "200":
          description: Safe command argument accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CmdiResponse"
        "422":
          description: Dangerous command argument rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /smoke/path-traversal/{path}:
    get:
      summary: Path traversal payload rejected via SafePath
      operationId: smokePathTraversal
      tags: [input-validation]
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Safe path accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathResponse"
        "422":
          description: Path traversal blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /smoke/xxe:
    post:
      summary: XXE entity expansion blocked via SecureXml
      operationId: smokeXxe
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
      responses:
        "200":
          description: Safe XML parsed
          content:
            application/json:
              schema:
                type: object
        "400":
          description: XXE payload rejected
  /smoke/deserialization:
    post:
      summary: Unsafe deserialization blocked via SecureJson
      operationId: smokeDeserialization
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeserializationRequest"
      responses:
        "200":
          description: Safe JSON deserialized
          content:
            application/json:
              schema:
                type: object
  /smoke/mass-assignment:
    post:
      summary: Mass assignment blocked via deny_unknown_fields
      operationId: smokeMassAssignment
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MassAssignmentRequest"
      responses:
        "200":
          description: Valid fields accepted
          content:
            application/json:
              schema:
                type: object
        "422":
          description: Unknown fields rejected
  /smoke/header-injection:
    post:
      summary: CRLF header injection blocked
      operationId: smokeHeaderInjection
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeaderInjectionRequest"
      responses:
        "200":
          description: Safe header value accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeaderInjectionResponse"
        "422":
          description: CRLF injection blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /smoke/unicode-bypass:
    post:
      summary: Unicode normalisation bypass check
      operationId: smokeUnicodeBypass
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnicodeBypassRequest"
      responses:
        "200":
          description: Unicode normalisation applied
          content:
            application/json:
              schema:
                type: object
  /smoke/body-bomb:
    post:
      summary: Oversized request body rejected
      operationId: smokeBodyBomb
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Body within limits
        "413":
          description: Body too large
  /smoke/deep-nesting:
    post:
      summary: Deeply nested JSON rejected
      operationId: smokeDeepNesting
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeepNestingRequest"
      responses:
        "200":
          description: Nesting within limits
        "422":
          description: Nesting depth exceeded
  /smoke/field-flood:
    post:
      summary: Excessive field count rejected
      operationId: smokeFieldFlood
      tags: [input-validation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Field count within limits
        "422":
          description: Too many fields

  # ── Output Encoding ───────────────────────────────────────────────────
...